requirements.txt
Flask>=2.2
gunicorn
requests
fpdf
reportlab
pillow
qrcode
numpy
matplotlib
APScheduler
pycryptodomex
